fixed the moto control

+++ smart-bridge/src/functionalities/waterflow-control-system/WaterflowContolSystem.h
@@ -8,10 +8,10 @@
 #include "logic/timer/Timer.h"
 #include "State.h"
 
-#define MINIMUM_SONAR_DISTANCE 0.02
-#define WL2_BOUND 1.5
-#define WL1_BOUND 3.00
-#define MAXIMUM_SONAR_DISTANCE 4.00 
+#define MINIMUM_SONAR_DISTANCE (float)0.02
+#define WL2_BOUND (float)1.5
+#define WL1_BOUND (float)3.00
+#define MAXIMUM_SONAR_DISTANCE (float)4.00 
 
 class WaterflowControlSystem{
     private:
